client: add support for overlayfs.v24
As of version 24, overlayfs requires workdir= mount option to allow
atomic operations (workdir and rootdir must reside on the same fs).

There is no straightforward and clean way to test the version of the
overlayfs module. It could be checked with modinfo and then compared to
a predefined set, but that would introduce an unnecessary maintenance
burden. This commit changes the overlay mounting so that it will first
try to mount overlays with the old style mount options (without workdir)
and if the mount fails, then with new style options.