client: in desktop sessions, use user's session device settings for xrandr
Previously, xrandr settings were applied from /etc/puavo/device.json,
which did not work for thinclients, because the device in which desktop
sessions are run is (TADAA!) desktopserver, not thinclient.

This commit changes the xrandr configuration of desktop sessions to use
the device entry from the user's session object instead.