install: cleanup temporary unfinished images before rdiff patching starts
This commit fixes an update problem in the following scenario:

  1. There is enough available diskspace in /images, .rdiff got downloaded
     succesfully and `rdiff patch` was started.

  2. Command `rdiff patch` unfortunately got interrupted (for example on reboot
     or shutdown) and a big, nearly finished, temporary image file was left
     behind filling /images.

  3. The available diskspace in /images is less than 1.33 * sizeof(ltsp.img).

  4. The next time the update was initiated, it failed due to lack of diskspace.

The problem could have been fixed at least in two different ways:

  1. When evaluating the required disk space, subtract the size of the
     unfinished image from the total.

  2. Remove unfinished images before evaluating the required disk space.

I ended up using method #2, because it was simpler and because `rdiff patch`
does not currently continue from previously unfinished image, because it writes
to stdout (instead of writing inplace to an image file) for pv to show the
progress stats. So, removing the previously unfinished temporary image before
patching does not affect the patching at all, it started from scratch anyway.